// apps/admin/src/lib/api/generated/models/adminIdentity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.6
 */

export interface AdminIdentity {
  csrf_token: string;
  email?: string | null;
  expires_at: number;
  issuer: string;
  name?: string | null;
  organization_id: string;
  roles: string[];
  subject: string;
}